requirements.txt
flask==3.0.3
torch==2.11.0+cu128
torchvision==0.26.0+cu128
Pillow==10.4.0
scikit-learn==1.5.2
kagglehub==0.3.3
